requirements.txt
# Core Framework
fastapi>=0.115.12
uvicorn[standard]>=0.24.0
pydantic>=2.11.5
pydantic-settings>=2.9.1

# Database
sqlalchemy>=2.0.41
alembic>=1.16.1
psycopg2-binary>=2.9.10
asyncpg>=0.30.0

# Redis & Caching
redis>=6.2.0

# Authentication & Security
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
cryptography>=44.0.0

# HTTP Clients
httpx>=0.28.1
aiofiles>=24.1.0

# Development & Testing
pytest>=8.3.5
pytest-asyncio>=1.0.0

# Environment
python-dotenv>=1.0.1
